requirements.txt
motmetrics>=1.1.0
numpy>=1.19.0
pandas>=1.0.0
opencv-python>=4.5.0
pyyaml>=5.4.0
prettytable>=0.7.0
scipy>=1.5.0
torch>=1.7.0
torchvision>=0.8.0
